refactor: create a base class for Swing-FX file choosers
